Fixed bugs in custom _strncmp function

+++ shell.h
@@ -26,7 +26,7 @@ size_t _strlen(const char *str);
 char *_strncpy(char *dest, const char *src, size_t n);
 char *_strtok(char *str, const char *delim);
 int _atoi(char *s);
-int _strncmp(const char *s1, const char *s2, int n);
+int _strncmp(const char *s1, const char *s2, size_t n);
 
 /** Memory util functions */
 void free2DArray(char **arr);
